Factor out test_helper module
Factor out a new test-only module containing helper functionality,
test_helper, and move the find_the_answer_fn() function into it. This
way, we will be able to reuse it from various tests instead of
duplicating the logic all over the place.
